+++ aem-intellij-core/src/main/kotlin/com/aemtools/completion/clientlib/provider/ClientlibDeclarationIncludeCompletionProvider.kt
@@ -1,10 +1,8 @@
 package com.aemtools.completion.clientlib.provider
 
 import com.aemtools.common.completion.lookupElement
-import com.aemtools.common.util.findParentByType
-import com.aemtools.common.util.getPsi
-import com.aemtools.common.util.relativeTo
-import com.aemtools.common.util.toPsiFile
+import com.aemtools.common.util.*
+import com.aemtools.index.ClientlibDeclarationIndexFacade
 import com.aemtools.lang.clientlib.CdLanguage
 import com.aemtools.lang.clientlib.psi.CdBasePath
 import com.aemtools.lang.clientlib.psi.CdInclude
@@ -14,7 +12,9 @@ import com.intellij.codeInsight.completion.CompletionParameters
 import com.intellij.codeInsight.completion.CompletionProvider
 import com.intellij.codeInsight.completion.CompletionResultSet
 import com.intellij.openapi.project.DumbAware
+import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiDirectory
+import com.intellij.psi.PsiFile
 import com.intellij.util.ProcessingContext
 import javax.swing.Icon
 
@@ -33,7 +33,7 @@ object ClientlibDeclarationIncludeCompletionProvider : CompletionProvider<Comple
     val currentElement = parameters.position.findParentByType(CdInclude::class.java)
         ?: return
     val basePath = currentElement.basePathElement()?.include?.text
-    val startDir = findStartDirectory(parameters.originalFile.containingDirectory, basePath)
+    val startDir = findStartDirectory(parameters.originalFile, basePath)
         ?: return
 
     val rawVariants = collectRawVariants(parameters, startDir, basePath)
@@ -68,7 +68,7 @@ object ClientlibDeclarationIncludeCompletionProvider : CompletionProvider<Comple
     var currentBasePath: String? = null
     elements.forEach {
       if (it is CdBasePath) {
-        currentBasePath = it.include?.text
+        currentBasePath = findStartDirectory(cdFile, it.include?.text)?.virtualFile?.path
       } else if (it is CdInclude) {
         if (currentBasePath == null) {
           result.add(CdImport(
@@ -77,7 +77,7 @@ object ClientlibDeclarationIncludeCompletionProvider : CompletionProvider<Comple
           ))
         } else {
           result.add(CdImport(
-              realPath = "$realPath/$currentBasePath/${it.text}",
+              realPath = "$currentBasePath/${it.text}",
               relativePath = it.text,
               basePath = currentBasePath
           ))
@@ -95,7 +95,7 @@ object ClientlibDeclarationIncludeCompletionProvider : CompletionProvider<Comple
     fun variantsCollector(dir: PsiDirectory): List<CdImport> {
       val result = ArrayList<CdImport>()
       dir.files.filter {
-        accept(currentFile.name, it.name)
+        ClientlibDeclarationIndexFacade.isMatchedByExtension(it, currentFile.name)
       }
           .forEach {
             result.add(CdImport(
@@ -111,14 +111,22 @@ object ClientlibDeclarationIncludeCompletionProvider : CompletionProvider<Comple
     return variantsCollector(startDir)
   }
 
-  private fun findStartDirectory(directory: PsiDirectory, basePath: String?): PsiDirectory? {
+  private fun findStartDirectory(currentFile: PsiFile, basePath: String?): PsiDirectory? {
     if (basePath == null) {
-      return directory
+      return currentFile.containingDirectory
     }
 
+    if (isAbsolutePath(basePath)) {
+      return getAbsoluteDirectoryPath(currentFile, basePath)?.toPsiDirectory(currentFile.project)
+    }
+
+    return getRelativeDirectoryPath(basePath, currentFile)
+  }
+
+  private fun getRelativeDirectoryPath(basePath: String, currentFile: PsiFile): PsiDirectory? {
+    var result: PsiDirectory? = currentFile.containingDirectory
     val parts = basePath.split("\\", "/")
 
-    var result: PsiDirectory? = directory
     parts.forEach {
       when (it) {
         ".." -> result = result?.parentDirectory
@@ -131,11 +139,21 @@ object ClientlibDeclarationIncludeCompletionProvider : CompletionProvider<Comple
     return result
   }
 
-  private fun accept(currentFileName: String, fileToCheck: String): Boolean = if (currentFileName == "js.txt") {
-    fileToCheck.endsWith(".js")
-  } else {
-    fileToCheck.endsWith(".css") || fileToCheck.endsWith(".less")
+  private fun getAbsoluteDirectoryPath(file: PsiFile, basePath: String): VirtualFile? {
+    return ClientlibDeclarationIndexFacade.findAllTypedFiles(file)
+        .map { Pair(it.parent, it.parent.path.normalizeToJcrRoot()) }
+        .filter { it.second.startsWith(basePath) }
+        .minByOrNull { it.second.replaceFirst(basePath, "").split("/").size }
+        ?.let {
+          var directory = it.first
+          while (directory.path.normalizeToJcrRoot() != basePath) {
+            directory = directory.parent
+          }
+          return directory
+        }
   }
+
+  private fun isAbsolutePath(basePath: String) = basePath.startsWith("/")
 
   /**
    * File import data class.