+++ aem-intellij-core/src/test/kotlin/com/aemtools/completion/clientlib/ClientlibCssTxtCompletionTest.kt
@@ -4,11 +4,13 @@ import com.aemtools.common.constant.const.JCR_ROOT
 import com.aemtools.test.completion.CompletionBaseLightTest
 
 /**
+ * Tests for [CdCompletionContributor] with css/less files.
+ *
  * @author Dmytro Primshyts
  */
 class ClientlibCssTxtCompletionTest : CompletionBaseLightTest(false) {
 
-  fun testBaseCompletionCss() = completionTest {
+  fun `test css completion`() = completionTest {
     addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
             $CARET
         """)
@@ -20,7 +22,7 @@ class ClientlibCssTxtCompletionTest : CompletionBaseLightTest(false) {
     ))
   }
 
-  fun testCompletionWithBasePathCss() = completionTest {
+  fun `test completion with #base path under containing file dir`() = completionTest {
     addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
             #base=css
             $CARET
@@ -33,7 +35,7 @@ class ClientlibCssTxtCompletionTest : CompletionBaseLightTest(false) {
     ))
   }
 
-  fun testFilterOutDuplicates() = completionTest {
+  fun `test filter out duplicates without any #base path`() = completionTest {
     addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
             first.css
             $CARET
@@ -44,7 +46,7 @@ class ClientlibCssTxtCompletionTest : CompletionBaseLightTest(false) {
     shouldNotContain(listOf("first.css"))
   }
 
-  fun testFilterOutDuplicatesWithBasePath() = completionTest {
+  fun `test filter out duplicates with #base path under containing file dir`() = completionTest {
     addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
             css/first.css
             #base=css
@@ -56,7 +58,7 @@ class ClientlibCssTxtCompletionTest : CompletionBaseLightTest(false) {
     shouldNotContain(listOf("first.css"))
   }
 
-  fun testFilterOutDuplicatesAfter() = completionTest {
+  fun `test filter out duplicates after caret without any #base path`() = completionTest {
     addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
           $CARET
           first.css
@@ -68,7 +70,7 @@ class ClientlibCssTxtCompletionTest : CompletionBaseLightTest(false) {
     shouldNotContain(listOf("first.css"))
   }
 
-  fun testFilterOutDuplicatesWithBasePathAfter() = completionTest {
+  fun `test filter out duplicates after caret with #base path under containing file dir`() = completionTest {
     addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
           $CARET
           #base=css
@@ -81,7 +83,7 @@ class ClientlibCssTxtCompletionTest : CompletionBaseLightTest(false) {
     shouldNotContain(listOf("css/first.css", "first.css"))
   }
 
-  fun testBasePathCompletion1() = completionTest {
+  fun `test #base path completion under containing file dir`() = completionTest {
     addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
             #base=$CARET
         """)
@@ -91,7 +93,7 @@ class ClientlibCssTxtCompletionTest : CompletionBaseLightTest(false) {
     shouldNotContain(listOf("js"))
   }
 
-  fun testBasePathCompletion2() = completionTest {
+  fun `test #base path completion under containing file dir with different level directories`() = completionTest {
     addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
             #base=$CARET
         """)
@@ -101,6 +103,303 @@ class ClientlibCssTxtCompletionTest : CompletionBaseLightTest(false) {
         "css",
         "css/inner"
     ))
+  }
+
+  fun `test absolute #base path completion starting with root level`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=/a$CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path2/clientlibs/css/second.less", "")
+    shouldContain(listOf(
+        "apps/components/path1/clientlibs/css",
+        "apps/components/path2/clientlibs/css"
+    ))
+  }
+
+  fun `test absolute #base path completion starting with 3rd level path`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=/apps/components/a$CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/zero.js", "")
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/inner/clientlibs/css/second.css", "")
+    addFile("$JCR_ROOT/apps/components/path2/inner/clientlibs/css/third.less", "")
+    shouldContain(listOf(
+        "path1/clientlibs/css",
+        "path1/inner/clientlibs/css",
+        "path2/inner/clientlibs/css"
+    ))
+  }
+
+  fun `test absolute #base path completion starting with 5rd level path without matched files by typed character`() =
+      completionTest {
+        addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=/apps/components/path1/clientlibs/a$CARET
+        """)
+        addFile("$JCR_ROOT/apps/components/zero.js", "")
+        addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+        shouldContain(emptyList())
+      }
+
+  fun `test absolute #base path completion starting with 5rd level path with matched files by typed character`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=/apps/components/path1/clientlibs/s$CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/zero.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    shouldContain(listOf("css"))
+  }
+
+  fun `test relative #base path completion with -1 path level`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=../a$CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/zero.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/inner/clientlibs/css/second.less", "")
+    addFile("$JCR_ROOT/apps/components/path2/inner/clientlibs/css/third.less", "")
+    shouldContain(emptyList())
+  }
+
+  fun `test relative #base path completion with -2 path levels more globally`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=../../a$CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/zero.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/inner/clientlibs/css/second.less", "")
+    addFile("$JCR_ROOT/apps/components/path2/inner/clientlibs/less/third.less", "")
+    addFile("$JCR_ROOT/apps/components/path2/inner/clientlibs/less/third.css", "")
+    shouldContain(listOf(
+        "path1/clientlibs/css",
+        "path1/inner/clientlibs/css",
+        "path2/inner/clientlibs/less"
+    ))
+  }
+
+  fun `test relative #base path completion with -3 path levels`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=../../../c$CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/zero.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/inner/clientlibs/css/second.css", "")
+    addFile("$JCR_ROOT/apps/components/path2/inner/clientlibs/css/third.less", "")
+    shouldContain(listOf(
+        "components",
+        "components/path1/clientlibs/css",
+        "components/path1/inner/clientlibs/css",
+        "components/path2/inner/clientlibs/css"
+    ))
+  }
+
+  fun `test relative #base path completion with -2 path levels`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/path1/inner/clientlibs/css.txt", """
+            #base=../../c$CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/zero.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/inner/clientlibs/css/second.css", "")
+    addFile("$JCR_ROOT/apps/components/path2/inner/clientlibs/css/third.css", "")
+    shouldContain(listOf(
+        "clientlibs/css",
+        "inner/clientlibs/css"
+    ))
+  }
+
+  fun `test completion with absolute #base path with 3 levels`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=/apps/components/path1
+            $CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path2/clientlibs/css/second.css", "")
+    shouldContain(listOf(
+        "clientlibs/css/first.css"
+    ))
+  }
+
+  fun `test completion with absolute #base path with 2 levels`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=/apps/components
+            $CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path2/clientlibs/css/second.less", "")
+    shouldContain(listOf(
+        "path1/clientlibs/css/first.css",
+        "path2/clientlibs/css/second.less"
+    ))
+  }
+
+  fun `test filter out duplicates defined under 1 absolute #base path`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=/apps/components/path1/clientlibs/css
+            first.css
+            $CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/second.css", "")
+    shouldContain(listOf("second.css"))
+    shouldNotContain(listOf("first.css"))
+  }
+
+  fun `test filter out duplicates with absolute #base path and another #base path above`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=/apps/components/path1/clientlibs/css
+            first.css
+            
+            #base=/apps/components
+            $CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path2/clientlibs/css/second.css", "")
+    shouldContain(listOf("path2/clientlibs/css/second.css"))
+    shouldNotContain(listOf("path1/clientlibs/css/first.css"))
+  }
+
+  fun `test filter out duplicates with absolute #base path and another #base path below`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=/apps/components
+            $CARET
+            
+            #base=/apps/components/path1/clientlibs/css
+            first.less
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.less", "")
+    addFile("$JCR_ROOT/apps/components/path2/clientlibs/css/second.css", "")
+    shouldContain(listOf("path2/clientlibs/css/second.css"))
+    shouldNotContain(listOf("path1/clientlibs/css/first.less"))
+  }
+
+  fun `test completion with relative #base path with -2 levels and 1 subdir`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=../../path1
+            $CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/inner/clientlibs/css/second.css", "")
+    shouldContain(listOf(
+        "clientlibs/css/first.css",
+        "inner/clientlibs/css/second.css"
+    ))
+  }
+
+  fun `test completion with relative #base path with -2 levels and 3 subdirs`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=../../path1/clientlibs/css
+            $CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path2/clientlibs/css/second.css", "")
+    shouldContain(listOf("first.css"))
+  }
+
+  fun `test filter out duplicates defined under 1 relative #base path`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=../../path1/clientlibs/css
+            first.css
+            $CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/second.css", "")
+    shouldContain(listOf("second.css"))
+    shouldNotContain(listOf("first.css"))
+  }
+
+  fun `test filter out duplicates with relative #base path and another #base path above`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=../../path1
+            clientlibs/css/first.css
+            
+            #base=../../path1
+            $CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/inner/clientlibs/css/second.less", "")
+    shouldNotContain(listOf("clientlibs/css/first.css"))
+    shouldContain(listOf("inner/clientlibs/css/second.less"))
+  }
+
+  fun `test filter out duplicates with relative #base path and another #base path below`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=../../path1
+            $CARET
+            
+            #base=../../path1
+            clientlibs/css/first.css
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/inner/clientlibs/css/second.css", "")
+    shouldNotContain(listOf("clientlibs/css/first.css"))
+    shouldContain(listOf("inner/clientlibs/css/second.css"))
+  }
+
+  fun `test filter out duplicates with relative #base path and another absolute #base path below`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=../../path1
+            $CARET
+            
+            #base=/apps/components/path1/clientlibs/css
+            first.css
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/inner/clientlibs/css/second.css", "")
+    shouldNotContain(listOf("clientlibs/css/first.css"))
+    shouldContain(listOf("inner/clientlibs/css/second.css"))
+  }
+
+  fun `test filter out duplicates with absolute #base path and another relative #base path below`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/comp/clientlibs/css.txt", """
+            #base=/apps/components/path1
+            $CARET
+            
+            #base=../../path1/clientlibs/css
+            first.css
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/inner/clientlibs/css/second.less", "")
+    shouldNotContain(listOf("clientlibs/css/first.css"))
+    shouldContain(listOf("inner/clientlibs/css/second.less"))
+  }
+
+  fun `test filter out duplicates with abs #base path and #base path under containing file dir`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css.txt", """
+            #base=/apps/components/path1
+            $CARET
+            
+            #base=css
+            first.css
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/inner/clientlibs/css/second.css", "")
+    shouldNotContain(listOf("clientlibs/css/first.css"))
+    shouldContain(listOf("inner/clientlibs/css/second.css"))
+  }
+
+  fun `test completion with dot #base path `() = completionTest {
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css.txt", """
+            #base=.
+            $CARET
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/inner/clientlibs/css/second.css", "")
+    shouldContain(listOf("css/first.css"))
+  }
+
+  fun `test filter out duplicates with abs #base path and dot #base path`() = completionTest {
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css.txt", """
+            #base=.
+            $CARET
+            
+            #base=css
+            first.css
+        """)
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/first.css", "")
+    addFile("$JCR_ROOT/apps/components/path1/clientlibs/css/second.css", "")
+    shouldNotContain(listOf("css/first.css"))
+    shouldContain(listOf("css/second.css"))
   }
 
 }